Uses approach dist/inj depth as image projection
Previously, the "approach distance" and "injection depth" were used as
the "distance" along the injection (d) axis. However, inutitively it
makes more sense as a user that these distances be as visualized in the
image. Aka these distances should be along the along the manipulator's
x-axis (assuming the manipulator's x-y plane is parallel with the image
plane and the x-d plane is perpendicular to the image plane).